+++ Samples~/EyeTrackingExamples/Scripts/OpenXRLocalRigConfig.cs
@@ -52,7 +52,7 @@ namespace MagicLeap.MRTK.Samples.EyeTracking
                     openXRController.name = openXRControllerPrefab.name;
 
                     // Re-initialize interaction mode detectors.
-                    InteractionModeManager modeManager = GameObject.FindObjectOfType<InteractionModeManager>();
+                    InteractionModeManager modeManager = Object.FindAnyObjectByType<InteractionModeManager>();
                     if (modeManager != null)
                     {
                         modeManager.InitializeInteractionModeDetectors();